requirements.txt
# ThemeHunter 依赖

# 数据采集
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
lxml>=4.9.0

# 数据处理
pyyaml>=6.0
python-dateutil>=2.8.0

# Web框架（用于Bot）
flask>=2.0.0

# 可选：异步处理
aiohttp>=3.8.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
